+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ### Fixed
 - Allow trailing comments for GitHub syntax
+- AlreadyDisposedException was thrown when closing a project
 
 ### Changed
 - Update template to [1.4.0](https://github.com/JetBrains/intellij-platform-plugin-template/releases/tag/v1.4.0)